# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: MindTab API
  description: API for MindTab — task tracking and micro-notes platform.
  version: 0.1.0
  contact:
    name: MindTab
    url: https://mindtab.in
servers:
  - url: /
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTask
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [Tasks]
      summary: Get a single task
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Task with project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      tags: [Tasks]
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTask
      tags: [Tasks]
      summary: Soft-delete a task
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Task deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks/count:
    get:
      operationId: countTasks
      tags: [Tasks]
      summary: Count tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Task count
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tasks/unassigned:
    get:
      operationId: listUnassignedTasks
      tags: [Tasks]
      summary: List tasks without a project
      responses:
        "200":
          description: Unassigned tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /tasks/archive-completed:
    post:
      operationId: archiveCompletedTasks
      tags: [Tasks]
      summary: Archive all completed tasks
      responses:
        "200":
          description: Completed tasks archived
          content:
            application/json:
              schema:
                type: object
                required: [success, count]
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Habits ─────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Parameters ───────────────────────────────────────────────────────
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ─── Reusable Responses ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ─── Schemas ──────────────────────────────────────────────────────────
  schemas:
...
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed, archived]
    TaskPriority:
      type: string
      enum: [priority_1, priority_2, priority_3, priority_4]
    TaskImpact:
      type: string
      enum: [low, medium, high]
...
    ProjectStatus:
      type: string
      enum: [active, paused, completed, archived]

    # ── Error ──────────────────────────────────────────────────────────
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
          description: Machine-readable error code

    # ── Auth ───────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, status, priority, impact, position, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        impact:
          $ref: "#/components/schemas/TaskImpact"
        position:
          type: integer
        projectId:
          type: string
          format: uuid
          nullable: true
        project:
          $ref: "#/components/schemas/ProjectSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 256
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        impact:
          $ref: "#/components/schemas/TaskImpact"
        position:
          type: integer
        projectId:
          type: string
          format: uuid
          nullable: true
        completedAt:
          type: string
          format: date-time
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 256
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        impact:
          $ref: "#/components/schemas/TaskImpact"
        position:
          type: integer
        projectId:
          type: string
          format: uuid
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true

    # ── Habit ──────────────────────────────────────────────────────────
...
    ProjectSummary:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProjectStatus"
...